A2A/requirements.txt
# A2A (Agent-to-Agent) example dependencies
# google-adk[a2a] = ADK with A2A support (to_a2a, RemoteA2aAgent)
# uvicorn       = ASGI server for remote agents and registry
# litellm       = Model connector (Azure, Gemini, etc.)
# python-dotenv = Load .env for credentials
# fastapi       = Registry server (registry_server.py)
google-adk[a2a]>=0.1.0
uvicorn>=0.30.0
litellm>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
